pyproject.toml
[tool.poetry]
name = "python"
version = "0.1.0"
description = ""
authors = ["umibe yasushi <umibe@otoso.jp>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
chromalog = "^1.0.5"
PySimpleGUI = "^4.49.0"
matplotlib = "^3.4.3"
pathlib = "^1.0.1"
pandas = "^1.3.3"
scipy = "^1.7.1"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
types-all = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
